pyproject.toml
[project]
name = "examen-bentoml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bentoml>=1.4.27",
    "dvc>=3.63.0",
    "dvc-s3>=3.2.2",
    "ipykernel>=7.0.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
